More logic moved into core classes
