+++ src/main/java/com/hotels/beeju/BeejuJUnitRule.java
@@ -51,7 +51,7 @@ abstract class BeejuJUnitRule extends ExternalResource {
    */
   protected void init() throws Throwable {
     metastoreLocation = temporaryFolder.newFolder("metastore");
-    core.setHiveVar(HiveConf.ConfVars.METASTOREWAREHOUSE, metastoreLocation.getAbsolutePath() );
+    core.setHiveVar(HiveConf.ConfVars.METASTOREWAREHOUSE, metastoreLocation.getAbsolutePath());
   }
 
   /**
@@ -69,7 +69,7 @@ abstract class BeejuJUnitRule extends ExternalResource {
     temporaryFolder.create();
     init();
     beforeTest();
-    createDatabase(core.databaseName());
+    createDatabase(databaseName());
   }
 
   /**
@@ -108,6 +108,14 @@ abstract class BeejuJUnitRule extends ExternalResource {
   }
 
   /**
+   * @return a copy of the {@link HiveConf} used to create the Hive Metastore database. This {@link HiveConf} should be
+   *         used by tests wishing to connect to the database.
+   */
+  public HiveConf conf() {
+    return core.conf();
+  }
+
+  /**
    * Create a new database with the specified name.
    *
    * @param databaseName Database name.
@@ -115,7 +123,7 @@ abstract class BeejuJUnitRule extends ExternalResource {
    */
   public void createDatabase(String databaseName) throws TException {
     File tempFolder = temporaryFolder.getRoot();
-    core.createDatabase(databaseName,tempFolder, core.conf());
+    core.createDatabase(databaseName,tempFolder, conf());
   }
 
 }
