+++ src/test/java/com/hotels/beeju/HiveServer2JUnitRuleTest.java
@@ -63,7 +63,7 @@ public class HiveServer2JUnitRuleTest {
   public void customProperties() {
     Map<String, String> conf = new HashMap<>();
     conf.put("my.custom.key", "my.custom.value");
-    HiveConf hiveConf = new HiveServer2JUnitRule(DATABASE, conf).core.conf();
+    HiveConf hiveConf = new HiveServer2JUnitRule(DATABASE, conf).conf();
     assertThat(hiveConf.get("my.custom.key"), is("my.custom.value"));
   }
